Renamed web/shared/middleware -> web/shared/middlewares
