+++ core/server/api/authentication.js
@@ -10,7 +10,7 @@ const Promise = require('bluebird'),
     urlService = require('../services/url'),
     localUtils = require('./utils'),
     models = require('../models'),
-    spamPrevention = require('../web/shared/middleware/api/spam-prevention'),
+    spamPrevention = require('../web/shared/middlewares/api/spam-prevention'),
     mailAPI = require('./mail'),
     settingsAPI = require('./settings'),
     tokenSecurity = {};
