+++ core/server/apps/private-blogging/lib/router.js
@@ -3,7 +3,7 @@ const path = require('path'),
     middleware = require('./middleware'),
     bodyParser = require('body-parser'),
     routing = require('../../../services/routing'),
-    brute = require('../../../web/shared/middleware/brute'),
+    brute = require('../../../web/shared/middlewares/brute'),
     templateName = 'private',
     privateRouter = express.Router();
