+++ core/server/services/auth/oauth.js
@@ -3,7 +3,7 @@ var oauth2orize = require('oauth2orize'),
     passport = require('passport'),
     models = require('../../models'),
     authUtils = require('./utils'),
-    spamPrevention = require('../../web/shared/middleware/api/spam-prevention'),
+    spamPrevention = require('../../web/shared/middlewares/api/spam-prevention'),
     common = require('../../lib/common'),
     oauthServer,
     oauth;
