+++ core/server/web/shared/middlewares/admin-redirects.js
