+++ core/server/web/shared/middlewares/api/cors.js
