+++ core/server/web/shared/middlewares/error-handler.js
