+++ core/server/web/shared/middlewares/frontend-client.js
