+++ core/server/web/shared/middlewares/ghost-locals.js
