+++ core/server/web/shared/middlewares/image/index.js
