+++ core/server/web/shared/middlewares/image/normalize.js
