+++ core/server/web/shared/middlewares/index.js
