+++ core/server/web/shared/middlewares/maintenance.js
