+++ core/server/web/shared/middlewares/pretty-urls.js
