+++ core/server/web/shared/middlewares/serve-favicon.js
