+++ core/server/web/shared/middlewares/serve-public-file.js
