+++ core/server/web/shared/middlewares/static-theme.js
