+++ core/server/web/shared/middlewares/url-redirects.js
