+++ core/server/web/shared/middlewares/validation/blog-icon.js
